Base layout creation
Criação de um layout básica para as páginas
