+++ index.html
@@ -7,7 +7,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
       rel="stylesheet"
-      href="https://fonts.googleapis.com/css2?family=Anybody:wght@500&family=Bubblegum+Sans&family=Bungee&family=Satisfy&family=Varela+Round&display=swap"
+      href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700&family=Anybody:wght@500&family=Bubblegum+Sans&family=Bungee&family=Satisfy&family=Varela+Round&display=swap"
     />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>HexAnalytics</title>
