+++ package.json
@@ -14,6 +14,7 @@
     "@amcharts/amcharts4-geodata": "^4.1.28",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.0",
+    "@fontsource/roboto": "^5.0.12",
     "@iconify/react": "^4.1.1",
     "@mui/icons-material": "^5.15.13",
     "@mui/material": "^5.15.13",
