+++ src/App.jsx
@@ -1,6 +1,6 @@
-import React from "react"
-import AppRoutes from "./Routes"
+import React from "react";
+import AppRoutes from "./Routes";
 
 export default function App() {
-	return <AppRoutes />
+  return <AppRoutes />
 }
