+++ src/components/Sidebar.jsx
@@ -1,0 +1,65 @@
+import * as React from 'react';
+import ListItemButton from '@mui/material/ListItemButton';
+import ListItemIcon from '@mui/material/ListItemIcon';
+import ListItemText from '@mui/material/ListItemText';
+import ListSubheader from '@mui/material/ListSubheader';
+import HomeRoundedIcon from '@mui/icons-material/HomeRounded';
+import DashboardRoundedIcon from '@mui/icons-material/DashboardRounded';
+import MapRoundedIcon from '@mui/icons-material/MapRounded';
+import FolderSharedRoundedIcon from '@mui/icons-material/FolderSharedRounded';
+import AssignmentIcon from '@mui/icons-material/Assignment';
+
+export const mainListItems = (
+  <React.Fragment>
+    <ListItemButton href="/">
+      <ListItemIcon>
+        <HomeRoundedIcon />
+      </ListItemIcon>
+      <ListItemText primary="Home" />
+    </ListItemButton>
+    <ListItemButton href="/dashboard">
+      <ListItemIcon>
+        <DashboardRoundedIcon />
+      </ListItemIcon>
+      <ListItemText primary="Dashboard" />
+    </ListItemButton>
+    <ListItemButton href="/">
+      <ListItemIcon>
+        <MapRoundedIcon/>
+      </ListItemIcon>
+      <ListItemText primary="Map" />
+    </ListItemButton>
+    <ListItemButton>
+      <ListItemIcon>
+        <FolderSharedRoundedIcon/>
+      </ListItemIcon>
+      <ListItemText primary="Your data" />
+    </ListItemButton>
+  </React.Fragment>
+);
+
+export const secondaryListItems = (
+  <React.Fragment>
+    <ListSubheader component="div" inset>
+      Saved reports
+    </ListSubheader>
+    <ListItemButton>
+      <ListItemIcon>
+        <AssignmentIcon />
+      </ListItemIcon>
+      <ListItemText primary="Current month" />
+    </ListItemButton>
+    <ListItemButton>
+      <ListItemIcon>
+        <AssignmentIcon />
+      </ListItemIcon>
+      <ListItemText primary="Last quarter" />
+    </ListItemButton>
+    <ListItemButton>
+      <ListItemIcon>
+        <AssignmentIcon />
+      </ListItemIcon>
+      <ListItemText primary="Year-end sale" />
+    </ListItemButton>
+  </React.Fragment>
+);
