+++ src/pages/Template.jsx
@@ -1,0 +1,9 @@
+import BaseLayout from "../layouts/BaseLayout"
+
+export default function Template() {
+  return (
+    <BaseLayout titulo="Título da página">
+      {/* {Conteúdo da página} */}
+    </BaseLayout>
+  );
+}
